+++ Codecademy_Vetoffice/settings.py
@@ -37,7 +37,10 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'vetoffice.apps.VetofficeConfig'
+    'vetoffice.apps.VetofficeConfig',
+    'employee',
+    'owner',
+    'patient'
 ]
 
 MIDDLEWARE = [
