+++ employee/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class EmployeeConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'employee'
