+++ owner/templates/owner/owner_delete_form.html
