+++ patient/templates/patient/breed_create_form.html
@@ -1,0 +1,11 @@
+{% extends './base.html' %}
+{% block content %}
+<h2>Register new Breed</h2>
+<form method="post">
+  <div>
+    {% csrf_token %}
+    {{ form.as_p }}
+    <input type="submit" value="Submit" />
+  </div>
+</form>
+{% endblock %}
