+++ patient/templates/patient/breed_list.html
@@ -1,0 +1,23 @@
+{% extends './base.html' %}
+{% block content %}
+<h2>Breed List</h2>
+<a href="{% url 'vetoffice:breedcreate' %}"><button class"btn">Add Breed</button></a>
+<table id="breeds">
+  <tr>
+    <th>Animal Type</th>
+    <th>Breed</th>
+    <th> </th>
+    <th> </th>
+  </tr>
+  <tr>
+    {% for breed in breed_list %}
+  </tr>
+  <tr>
+    <td>{{ breed.animal }}</td>
+    <td>{{ breed.breed }}</td>
+    <td><a href="{% url 'vetoffice:breedupdate' breed.id %}"><button>Edit</button></a></td>
+    <td><a href="{% url 'vetoffice:breeddelete' breed.id %}"><button>Delete</button></a></td>
+  </tr>
+  {% endfor %}
+</table>
+{% endblock %}
