+++ patient/templates/patient/patient_delete_form.html
