+++ vetoffice/urls.py
@@ -1,7 +1,9 @@
 from django.urls import path
 #from django.contrib import admin
 
-from . import views
+from owner import views
+from patient import views
+from employee import views
 
 app_name = 'vetoffice'
